标签删除功能补偿

+++ bbs-springboot/bbs-rest/src/main/java/com/liang/bbs/rest/controller/LabelController.java
@@ -8,9 +8,11 @@ import com.liang.bbs.rest.config.login.NoNeedLogin;
 import com.liang.bbs.rest.config.swagger.ApiVersion;
 import com.liang.bbs.rest.config.swagger.ApiVersionConstant;
 import com.liang.bbs.rest.utils.FileLengthUtils;
+import com.liang.nansheng.common.auth.RoleSsoDTO;
 import com.liang.nansheng.common.auth.UserContextUtils;
 import com.liang.nansheng.common.auth.UserSsoDTO;
 import com.liang.nansheng.common.enums.ResponseCode;
+import com.liang.nansheng.common.enums.RoleGradeEnum;
 import com.liang.nansheng.common.utils.CommonUtils;
 import com.liang.nansheng.common.web.basic.ResponseResult;
 import com.liang.nansheng.common.web.exception.BusinessException;
@@ -19,10 +21,18 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.dubbo.config.annotation.DubboReference;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RestController
@@ -44,15 +54,14 @@ public class LabelController {
     }
 
     @PostMapping("/uploadLabelLogo")
-    @Operation(summary = "上传标签logo")
+    @Operation(summary = "上传标签 logo")
     @ApiVersion(group = ApiVersionConstant.V_300)
     public ResponseResult<String> uploadLabelLogo(@RequestParam(value = "logo", required = false) MultipartFile logo) throws IOException {
         if (fileLengthUtils.isFileNotTooBig(logo.getBytes())) {
             return ResponseResult.success(labelService.uploadLabelLogo(logo.getBytes(), logo.getOriginalFilename()));
-        } else {
-            throw BusinessException.build(ResponseCode.EXCEED_THE_MAX, "请上传不超过 " +
-                    CommonUtils.byteConversion(fileLengthUtils.getFileMaxLength()) + " 的图片!");
         }
+        throw BusinessException.build(ResponseCode.EXCEED_THE_MAX,
+                "请上传不超过 " + CommonUtils.byteConversion(fileLengthUtils.getFileMaxLength()) + " 的图片");
     }
 
     @PostMapping("update")
@@ -75,7 +84,14 @@ public class LabelController {
     @Operation(summary = "删除标签")
     @ApiVersion(group = ApiVersionConstant.V_300)
     public ResponseResult<Boolean> delete(@PathVariable Integer id) {
+        UserSsoDTO currentUser = UserContextUtils.currentUser();
+        List<String> grades = currentUser.getRoles().stream()
+                .map(RoleSsoDTO::getGrade)
+                .distinct()
+                .collect(Collectors.toList());
+        if (!grades.contains(RoleGradeEnum.NS_SUPER_ADMIN_ROLE.name())) {
+            throw BusinessException.build(ResponseCode.URL_ACCESS_REFUSED);
+        }
         return ResponseResult.success(labelService.delete(id));
     }
-
 }
